==> 1_Introduction - Business_Understanding.ipynb <==
# UVA.Capstone1_webscrapper
### CS 5010, Summer 2020

* Alfred Hull, adh8mb
* Farhan Kanani, fk3ak
* Dara Maguire, dkm2bd
* Francisco Estrada, fge8tj
* Niraja Bohidar, nb4dg

### Phase I. Business Understanding.

### Background

As of 2019, we have found ourselves at a new precipice in entering the Fourth Industrial Revolution: The Real-Time Enterprise! Just like the intersections of the past: First in 1784, Water and steam; Second in 1870, First conveyor belt; and the Third in 1969, Electronics and information technology, this pause will bring with it both challenges and new opportunities. Data stewards in the discipline have realized associative costs for data storage has expounded this issue while driving talent gaps. As organizations continue to ingest voluminous amounts of data, they must become more tactical with the data they are creating and using to sustain or improve Return on Investment (ROI). More so, the approach of advancing insight through analyses of mountains of data must be clear and consistent so that results are impactful, insightful, but more so reproducible. 

Given the current shortfall of talent proportionate to the opportunities seen in mining volumes of data and the time-consuming aspect in discovering the best possible options, it is ever more critical that machine intelligence and automation play a progressive role within organizations. An extremely effective way of optimizing selection in a sea of choice is by deploying recommender systems. Recommender Systems are capable of predicting the future inclination of a set of items for an individual, team, or even entire organization, then optimizing selection based on commonalities of historical transactions. One fundamental value propositions of a recommender system in modern society is that it enables people to maximize choice/selection due to an onslaught of varieties with the prevalence of the Internet and a decreasing cost of data storage. Imagine going into your local library and searching for a cookbook. The librarian tells you that there is an entire wing of the library dedicated to cookbooks: have fun. While the information collected in the library may sound impressive, the hunt for what you may be looking for will undoubtedly be exhausting. If there was a way to provide insight about your needs given your comfort level of sharing your data, learning algorithms could help match your ideas with others who are similar then prescribe analogs that may better suit your needs. 

This project progresses logically through a recommender analysis utilizing the Cross-Industry Standard Process for Data Mining (CRISP-DM). The aim is enabling data stewards and stakeholders to clearly understand what, when, where, why, and how the Data Science work proceeds. The phases include Business understanding, Data Understanding, Data Preparation, Modeling & Application Development, Evaluation, Model Deployment.

### Who might care:
Maintenance Managers, Operations Managers, Knowledge Management Teams, Human Resource Departments in organizations, and especially tech organizations such as Microsoft, Amazon, Google, Apple, and individual agencies within the Department of Defense, etc. can use such a model to better market products and or materials to customers.

### Set-Objective:

The objective here is to build a Collaborative Filtering Systems based Recommender Systems that will predict the rating of a movie for a user based on the history of ratings the user applied to other films.

As users log into an entertainment platform and want to watch a movie: the recommendation system will need to suggest the movie rating based on the ratings the user has provided for other movies before. The system will also compare the ratings of other users who have watched similar videos to enhance the pushed selections. The business reasons to do this are many, and the important ones are:

1.	Relevancy. Movie ratings must be relevant for the user, taking into consideration the user's past ratings and interests. For instance, if the user ranks action movies high and drama movies low, the dichotomy in the user’s selection must be reflected here.

2.	Customer Retention. The system must put in its best efforts to keep its customers. For example, if the system recommends a movie to a user, and the customer liked it, they may continue coming back for more content. 

### Cost and benefits:

While the benefits of recommendation systems are exponentially impactful to an organization, fundamentally shifting the labor costs associated with marketing re-targeting efforts by automation. The system development is not natural to guesstimate from a top-level view. It requires a personalized approach and thorough understanding and analysis of a client’s business processes, set-objective, and data. Roughly the cost for this would range from 25,000.00 to 45,000.00. One way to build a rough order magnitude to factor costs is by looking at such a project through four key areas: 

1.	Analysis and rough order magnitude (ROM) estimation – 5,000.00 (40 hours burdened)
-	Business Understanding. Preliminary to the project start date, it is critical to set up set-objectives (business-understanding). The primary goal of this phase is to work with business analysts to conduct a feasibility study. The work on every project starts with this analysis. 
-	Data Understanding. Here we analyze currently used metrics, data assets, and customer set-objectives and business processes. 
-	Data Preparation. In this phase, the team defines growth points, determines the appropriate technological stack, timelines, and budgets, while developing all corresponding documentation with wrangled data.

2.	Prototype development – 5,000.00 - 10,000.00 (40 to 90 hours burdened)
-	Data Modeling. During this phase, we develop a first-go prototype of the recommendation system based on the data gathered during the previous step (Data Preparation) to test the hypothesis and show its efficiency. The Business and Data Understanding Phases are critical to calculating any final costs of AI-powered recommender development.

3.	Minimal Viable Product (MVP) development (with prototype included) – 10,000.00 - 15,000.00 (90 to 140 hours  burdened)
- Model Evaluation. The previous step was just a simple prototype only and not the alpha version of the recommendation system. Here we will evaluate the alpha version for deployment.

The main difference between Prototype Development and Minimum Viable Product (MVP) is as follows. The first proves the concept or idea –to extract insights from the data, and according to those insights generate suggestions and provide recommendations; MVP, on the other hand, is the original alpha version (feature testing) of a product that solves the set-objective.

4.	Deployment and Release – 5,000.00 – 15,000.00 (40 to 140 hours fully burdened)
-	Deployment & Conclusion. During the last phase, improve the MVP of the recommendation engine to match the customers' needs and integrate it within any existing infrastructures. 

The final release can, in some cases, be delayed or even postponed according to various factors. Often we choose agile software development and continuous integration/continuous development (CI/CD) and prefer to launch the recommender prototype, improving it over time while in production.


### Constraints, Limitations, and Assumptions (CLAs):

The most common sorts of recommendation systems are content-based and collaborative filtering. In collaborative filtering, the behavior of a cluster of users form recommendations to other users. The actual recommendations predicate the preference of other users. For example, recommending a movie to a consumer because their friend also liked the movie. There are two sorts of recommender engines: Memory-based and Model-based  methods. The simplest to build and deploy the three are memory-based techniques. The advantage of memory-based technologies is that they're simple to implement, and therefore the resulting recommendations are often both easy to communicate and evaluate.

__Memory-Based methods:__

1.	User-based collaborative filtering suggests products to a consumer based on the very fact that other users like the products with a high degree of commonalities.

2.	Item-based collaborative filtering Systems that identify similar items that supported users’ previous ratings. For example, Consumer A, B, and C gave thumbs-up rating to books X and Y. The next user D buys book Y, they get a recommendation to purchase book X. This is because the model identifies book X and Y as being similarly supported by the ratings of users A, B, and C.

3.	Content-based methods. Use metadata like genre, producer, actor, musician to recommend items say movies or music. For example, you were supporting Infinity War that featured Robert Downey Junior because someone watched and liked Iron Man. Similarly, you'll get music recommendations from individual artists because you liked their music. Content-based systems support the thought that if you wanted a particular item, you're presumably to love something that's almost like it.  

__Model-based methods:__

Are established using data mining, machine learning algorithms to predict users’ ratings on unrated items. During this approach, techniques such as dimensionality reduction improve model accuracy. Examples of such model-based methods include decision trees, rule-based models, Bayesian methods, and latent factor models.
