==> 3_Beyond_usual_specs - Data_Preparation.ipynb <==
# imports
import csv
import requests
import pandas as pd
from bs4 import BeautifulSoup as bs
import numpy as np
import re
import string
from IPython.core.display import HTML

# settings
pd.set_option('display.max_rows', 2000)
pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

# data viz imports
%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns

# get data
# read in covid19_states.csv
df1_covidstatecases = pd.read_csv("data/covid19_states_cases.csv", encoding='utf-8')
# read in covid19_states_humidity.csv
df2_statehumidity = pd.read_csv("data/covid19_states_humidity.csv", encoding='utf-8')

# glimpse at data to ensure it read in appropriately
print(df1_covidstatecases.head())
print(df2_statehumidity.head())

# apply a first round of data wrangling: text cleaning techniques
def clean_text_round1(text):
    '''Make text lowercase, remove text in square brackets, remove punctuation and remove words containing numbers.'''
    text = text.lower()
    text = re.sub('\[.*?\]', '', text)
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    text = re.sub('\w*\d\w*', '', text)
    text = re.sub('[‘’“”…]', '', text)
    text = re.sub('\n', '', text)
    return text

round1 = lambda x: clean_text_round1(x)

# lets wrangle USAState variable in the humidity dataframe
df2_statehumidity['USAState'] = pd.DataFrame(df2_statehumidity['USAState'].apply(round1))
df2_statehumidity['USAState'] = df2_statehumidity['USAState'].str.strip()

# lets wrangle USAState variable in the humidity dataframe & strip out leading and lagging whitespaces
df1_covidstatecases['USAState'] = pd.DataFrame(df1_covidstatecases['USAState'].apply(round1))
df1_covidstatecases['USAState'] = df1_covidstatecases['USAState'].str.strip()

# here we look at value fields for primary keys to see if the cleansing function worked
frames1 = [df1_covidstatecases['USAState'], df2_statehumidity['USAState']]
result1 = pd.concat(frames1).sort_values()
result1

# merge wrangled dataframes
merged = pd.merge(df1_covidstatecases,df2_statehumidity, how='right', on='USAState')

# select only neccessary column headers from merged file
df_cleanmerge = merged[['USAState', 'TotalCases', 'NewCases', 'TotalDeaths', 'NewDeaths', 'ActiveCases', 'Average Humidity', 'Population']]
df_cleanmerge.head()

# replace nan with zeroes
df_cleanmerge = df_cleanmerge.fillna(0)

# write clean merged file to .csv: df_cleanmerge
df_cleanmerge.to_csv("data\covid19_states_humidityandcases.csv", encoding='utf-8', index=False)
